pyproject.toml
[project]
name = "automations"
version = "0.1.0"
description = "Automation control panel with CRON scheduling support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0", "loguru>=0.7.0"]

[project.scripts]
automation-register = "scripts.register:main"
automation-run = "scripts.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
